dbeaver/pro#7023 refactor: move grid search logic to a separate plugin
