+++ webapp/common-react/@dbeaver/react-data-grid/src/BaseCell.tsx
@@ -1,6 +1,6 @@
 /*
  * CloudBeaver - Cloud Database Manager
- * Copyright (C) 2020-2025 DBeaver Corp and others
+ * Copyright (C) 2020-2026 DBeaver Corp and others
  *
  * Licensed under the Apache License, Version 2.0.
  * you may not use this file except in compliance with the License.
@@ -8,6 +8,7 @@
 
 import { memo, use, useMemo } from 'react';
 import { Cell, type CellRendererProps } from 'react-data-grid';
+import { clsx } from '@dbeaver/ui-kit';
 import { DataGridCellContext, type IDataGridCellRenderer } from './DataGridCellContext.js';
 import { createCellMouseEvent } from './eventUtils.js';
 import { DataGridCellInnerContext, type IDataGridCellInnerContext } from './DataGridCellInnerContext.js';
@@ -21,6 +22,7 @@ export const BaseCell = memo(function BaseCell<TRow, TSummaryRow>(props: CellRen
   const dataColIdx = dndContext.getDataColIdxByKey(props.column.key);
   const rowIdx = props.rowIdx;
   const tooltip = useGridReactiveValue(cellContext?.cellTooltip, rowIdx, dataColIdx);
+  const customClassName = useGridReactiveValue(cellContext?.getCellClassName, rowIdx, dataColIdx);
 
   function handleClick(event: React.MouseEvent<HTMLDivElement>) {
     props.onCellClick?.(
@@ -85,10 +87,11 @@ export const BaseCell = memo(function BaseCell<TRow, TSummaryRow>(props: CellRen
 
   const renderDefaultCell = useMemo<IDataGridCellRenderer>(
     () =>
-      ({ onClick, onDoubleClick, onContextMenu, ...rest }) => (
+      ({ onClick, onDoubleClick, onContextMenu, className, ...rest }) => (
         <Cell
           {...props}
           title={tooltip}
+          className={clsx(className, customClassName)}
           onCellClick={(_, event) => onClick?.(event)}
           onCellDoubleClick={(_, event) => onDoubleClick?.(event)}
           onCellContextMenu={(_, event) => onContextMenu?.(event)}
@@ -96,11 +99,15 @@ export const BaseCell = memo(function BaseCell<TRow, TSummaryRow>(props: CellRen
           isCellSelected={props.isCellSelected || rest.isFocused || false}
         />
       ),
-    [...Object.values(props), tooltip],
+    [...Object.values(props), tooltip, customClassName],
   );
 
   const cellElement = useGridReactiveValue(cellContext?.cellElement, rowIdx, dataColIdx, mappedProps, renderDefaultCell);
 
   const innerCellContext = useMemo<IDataGridCellInnerContext>(() => ({ isFocused: props.isCellSelected }), [props.isCellSelected]);
-  return <DataGridCellInnerContext value={innerCellContext}>{cellElement ?? <Cell title={tooltip} {...props} />}</DataGridCellInnerContext>;
+  return (
+    <DataGridCellInnerContext value={innerCellContext}>
+      {cellElement ?? <Cell title={tooltip} className={customClassName} {...props} />}
+    </DataGridCellInnerContext>
+  );
 }) as <TRow, TSummaryRow>(props: CellRendererProps<TRow, TSummaryRow>) => React.ReactNode;
