+++ webapp/common-react/@dbeaver/react-data-grid/src/DataGridCellContext.ts
@@ -25,6 +25,7 @@ export interface IDataGridCellContext {
   cell?: IGridReactiveValue<React.ReactNode, [rowIdx: number, colIdx: number]>;
   cellText?: IGridReactiveValue<string, [rowIdx: number, colIdx: number]>;
   cellTooltip?: IGridReactiveValue<string, [rowIdx: number, colIdx: number]>;
+  getCellClassName?: IGridReactiveValue<string | undefined, [rowIdx: number, colIdx: number]>;
   getCellEditable?: (rowIdx: number, colIdx: number) => boolean;
   onCellChange?: (rowIdx: number, colIdx: number, value: any) => void;
 }
