+++ webapp/packages/plugin-data-spreadsheet-new/src/DataGrid/DataGridContext.ts
@@ -1,14 +1,16 @@
 /*
  * CloudBeaver - Cloud Database Manager
- * Copyright (C) 2020-2025 DBeaver Corp and others
+ * Copyright (C) 2020-2026 DBeaver Corp and others
  *
  * Licensed under the Apache License, Version 2.0.
  * you may not use this file except in compliance with the License.
  */
 import { createContext } from 'react';
 
-import type { DataGridRef } from '@cloudbeaver/plugin-data-grid';
+
+import type {  DataGridRef } from '@cloudbeaver/plugin-data-grid';
 import type { IDatabaseDataModel, IDataTableActions } from '@cloudbeaver/plugin-data-viewer';
+import type { IGridSearchState } from '@cloudbeaver/plugin-grid-search';
 
 export interface IColumnResizeInfo {
   column: number;
@@ -23,6 +25,7 @@ export interface IDataGridContext {
   isGridInFocus: () => boolean;
   getDataGridApi: () => DataGridRef | null;
   focus: () => void;
+  searchState: IGridSearchState;
 }
 
 export const DataGridContext = createContext<IDataGridContext>(undefined as any);
