+++ webapp/packages/plugin-data-spreadsheet-new/tsconfig.json
@@ -69,6 +69,9 @@
     },
     {
       "path": "../plugin-data-viewer-conditional-formatting"
+    },
+    {
+      "path": "../plugin-grid-search"
     }
   ],
   "include": [
