+++ webapp/packages/plugin-grid-search/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@cloudbeaver/plugin-data-spreadsheet-new",
+  "name": "@cloudbeaver/plugin-grid-search",
   "type": "module",
   "sideEffects": [
     "./lib/module.js",
@@ -24,24 +24,9 @@
   },
   "dependencies": {
     "@cloudbeaver/core-blocks": "workspace:*",
-    "@cloudbeaver/core-browser": "workspace:*",
-    "@cloudbeaver/core-data-context": "workspace:*",
     "@cloudbeaver/core-di": "workspace:*",
-    "@cloudbeaver/core-dialogs": "workspace:*",
-    "@cloudbeaver/core-events": "workspace:*",
-    "@cloudbeaver/core-localization": "workspace:*",
-    "@cloudbeaver/core-sdk": "workspace:*",
-    "@cloudbeaver/core-settings": "workspace:*",
-    "@cloudbeaver/core-ui": "workspace:*",
     "@cloudbeaver/core-utils": "workspace:*",
-    "@cloudbeaver/core-view": "workspace:*",
-    "@cloudbeaver/plugin-data-grid": "workspace:*",
-    "@cloudbeaver/plugin-data-viewer": "workspace:*",
-    "@cloudbeaver/plugin-data-viewer-conditional-formatting": "workspace:*",
-    "@cloudbeaver/plugin-grid-search": "workspace:^",
-    "@dbeaver/js-helpers": "workspace:*",
-    "@dbeaver/result-set-api": "workspace:*",
-    "@dbeaver/ui-kit": "workspace:*",
+    "@dbeaver/ui-kit": "workspace:^",
     "mobx": "^6",
     "mobx-react-lite": "^4",
     "react": "^19",
@@ -51,13 +36,10 @@
   "devDependencies": {
     "@cloudbeaver/core-cli": "workspace:*",
     "@cloudbeaver/tsconfig": "workspace:*",
-    "@dbeaver/cli": "workspace:*",
-    "@dbeaver/react-tests": "workspace:*",
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "rimraf": "^6",
     "typescript": "^5",
-    "typescript-plugin-css-modules": "^5",
-    "vitest": "^4"
+    "typescript-plugin-css-modules": "^5"
   }
 }
