+++ webapp/packages/plugin-grid-search/src/PluginGridSearchBootstrap.ts
@@ -1,0 +1,19 @@
+/*
+ * CloudBeaver - Cloud Database Manager
+ * Copyright (C) 2020-2026 DBeaver Corp and others
+ *
+ * Licensed under the Apache License, Version 2.0.
+ * you may not use this file except in compliance with the License.
+ */
+
+import { Bootstrap, injectable } from '@cloudbeaver/core-di';
+
+@injectable()
+export class PluginGridSearchBootstrap extends Bootstrap {
+  constructor() {
+    super();
+  }
+
+  override register() {
+  }
+}
