+++ webapp/packages/plugin-grid-search/src/index.ts
@@ -1,0 +1,15 @@
+/*
+ * CloudBeaver - Cloud Database Manager
+ * Copyright (C) 2020-2026 DBeaver Corp and others
+ *
+ * Licensed under the Apache License, Version 2.0.
+ * you may not use this file except in compliance with the License.
+ */
+
+import './module.js';
+
+export * from './GridSearch/IGridSearchAdapter.js';
+export * from './GridSearch/GridSearchStore.js';
+export * from './GridSearch/GridSearchService.js';
+export * from './GridSearch/useGridSearch.js';
+export * from './GridSearch/GridSearchPanel.js';
