+++ webapp/packages/plugin-set-common/src/index.ts
@@ -109,6 +109,7 @@ import pluginCodemirror6 from '@cloudbeaver/plugin-codemirror6/module';
 import pluginAsyncTaskConfirmation from '@cloudbeaver/plugin-async-task-confirmation/module';
 import pluginSqlAsyncTaskConfirmation from '@cloudbeaver/plugin-sql-async-task-confirmation/module';
 import pluginDataViewerConditionalFormatting from '@cloudbeaver/plugin-data-viewer-conditional-formatting/module';
+import pluginGridSearch from '@cloudbeaver/plugin-grid-search/module';
 
 const core = [
   coreRouting, // important, should be first because the router starts in load phase first after all plugins register phase
@@ -217,4 +218,5 @@ export const commonSet = [
   pluginAsyncTaskConfirmation,
   pluginSqlAsyncTaskConfirmation,
   pluginDataViewerConditionalFormatting,
+  pluginGridSearch,
 ];
