+++ webapp/packages/plugin-set-common/tsconfig.json
@@ -191,6 +191,9 @@
       "path": "../plugin-gis-viewer"
     },
     {
+      "path": "../plugin-grid-search"
+    },
+    {
       "path": "../plugin-help"
     },
     {
